Move header files in the same folder than Go files
