+++ README.md
@@ -39,7 +39,7 @@ sections (`SEC("...")`). The following are currently supported:
 * `uretprobe/...`
 * `xdp/...`
 
-Map definitions must correspond to `bpf_map_def` from [the elf package](https://github.com/iovisor/gobpf/blob/master/elf/include/bpf_map.h).
+Map definitions must correspond to `bpf_map_def` from [the elf package](https://github.com/iovisor/gobpf/blob/master/elf/bpf_map.h).
 Otherwise, you will encounter an error like `only one map with size 280 bytes allowed per section (check bpf_map_def)`.
 
 The [Cilium](https://github.com/cilium/cilium) BPF docs contain helpful info
